Prepare future dl4j version updates and fix missing ndArray duplication
* Use slice methods instead of get/putRow to support later 3d and 4d
arrays
* Add missing dup() for ndArrays in AdversaryTrainingExample creation to
avoid overwriting ndArrays

+++ alpha-zero-adversary-learning/alpha-zero-learning-main/src/test/java/ch/evolutionsoft/rl/alphazero/AdversaryLearningTest.java
@@ -8,13 +8,17 @@ import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
+import org.apache.commons.math3.distribution.EnumeratedIntegerDistribution;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.factory.Nd4j;
 
+import cc.mallet.types.Dirichlet;
 import ch.evolutionsoft.rl.AdversaryLearningConfiguration;
 import ch.evolutionsoft.rl.AdversaryLearningSharedHelper;
 import ch.evolutionsoft.rl.AdversaryTrainingExample;
@@ -22,6 +26,7 @@ import ch.evolutionsoft.rl.Game;
 
 class AdversaryLearningTest {
 
+  public static final double DIRICHLET_WEIGHT = 0.5;
   public static final String TEST_TRAIN_EXAMPLES = "testTrainExamples.obj";
   public static final String TEST_TRAIN_EXAMPLES_VALUES = "testTrainExamplesValues.obj";
   AdversaryLearning adversaryLearning;
@@ -41,9 +46,10 @@ class AdversaryLearningTest {
   @AfterEach
   void deleteTestTrainExamples() throws IOException {
 
-    Files.delete(Paths.get(TEST_TRAIN_EXAMPLES));
-    Files.delete(Paths.get(TEST_TRAIN_EXAMPLES_VALUES));
-    
+    if (Files.exists(Paths.get(TEST_TRAIN_EXAMPLES))) {
+      Files.delete(Paths.get(TEST_TRAIN_EXAMPLES));
+      Files.delete(Paths.get(TEST_TRAIN_EXAMPLES_VALUES));
+    }
   }
 
   @Test
